Add graphics, and a few other improvements
This one's a bit of a mess because I wanted to get graphics done
so I could look at the results of other changes and work out what's
going wrong.

The -g option can now be used to enable graphics, which displays
some data relating to the electron distribution (that is kind of
specific to the callibration experiments I'm running with hydrogen
atoms in 3D, because it's easier to hardcode it).

A bias was eliminated from the energy estimator. It's now subject
to population control bias, but that's much smaller than the one
there previously was.

An ansatz that's exact for 3D hydrogen (and probably better than
the near-cusps-only one for general 3D systems) was added, and a
command line option to select the ansatz was added so I don't have
to keep hardcoding those changes.